Moved SleepDialog. Moved GameOutroManager.
