+++ src/engine/core/managers/outro/GameOutroManager.ts
