+++ src/engine/core/managers/outro/index.ts
@@ -1,0 +1,1 @@
+export * from "@/engine/core/managers/outro/GameOutroManager";
