+++ src/engine/core/managers/sleep/SleepConfig.ts
@@ -1,4 +1,4 @@
-import { SleepDialog } from "@/engine/core/ui/interaction/SleepDialog";
+import { SleepDialog } from "@/engine/core/ui/game/SleepDialog";
 import { Optional } from "@/engine/lib/types";
 
 export const sleepConfig = {
