+++ src/engine/core/managers/sleep/SleepManager.ts
@@ -8,7 +8,7 @@ import { sleepConfig } from "@/engine/core/managers/sleep/SleepConfig";
 import { surgeConfig } from "@/engine/core/managers/surge/SurgeConfig";
 import { SurgeManager } from "@/engine/core/managers/surge/SurgeManager";
 import { WeatherManager } from "@/engine/core/managers/weather/WeatherManager";
-import { SleepDialog } from "@/engine/core/ui/interaction/SleepDialog";
+import { SleepDialog } from "@/engine/core/ui/game/SleepDialog";
 import { executeConsoleCommand, getConsoleFloatCommand } from "@/engine/core/utils/console";
 import { disableInfoPortion, giveInfoPortion } from "@/engine/core/utils/info_portion";
 import { LuaLogger } from "@/engine/core/utils/logging";
